Separate files for linear solvers base classes.
