+++ resolve/LinSolverDirectCpuILU0.hpp
@@ -6,8 +6,9 @@
  * 
  */
 #pragma once
+
 #include "Common.hpp"
-#include "LinSolver.hpp"
+#include <resolve/LinSolverDirect.hpp>
 #include <resolve/MemoryUtils.hpp>
 
 namespace ReSolve
