+++ resolve/LinSolverDirectSerialILU0.hpp
@@ -1,9 +1,10 @@
 
 #pragma once
+
 #include "Common.hpp"
-#include "LinSolver.hpp"
 #include <resolve/MemoryUtils.hpp>
 #include <resolve/workspace/LinAlgWorkspace.hpp>
+#include <resolve/LinSolverDirect.hpp>
 
 namespace ReSolve 
 {
